tmt: add support for tmt framework
This adds support for running Zeppelin in a virtual machine by taking
advantage of the inbuilt virtual machine provisioning mechanisms in tmt
(test management tool).

This commit adds a simple test plan for tmt to execute which solely runs
the Zeppelin process and copies the data outputs back to the host
machine on completion.