prepare-osbuild: add support for rhel9 build hosts
Not all of the copr repositories have builds for all supported distros.
Add some conditional statements to fallback on epel-9 chroot if a rhel
host is detected to make a best attempt at installation.

The osbuild copr seems to select the centos stream 8 chroot on rhel9 so
this commit adds a conditional to force epel-9 to fix that.