roles: add prepare-distro-configs
Enable injecting/updating the distro config files in the sample-images
project. Allows defining of a list of dictionaries in config.yaml under
the variable "distros", which contain the contents to inject into the
distro configs (identified by the variable distro_name).

If the distro config already exists in the sample-images project, the
new values defined in config.yaml are merged on top of the existing
values.