roles: add run-make-osbuildvm
Add separate role to build the osbuildvm images. This commit adds the
role definitions, the steps to the local and Hetner playbooks, and
updates the config.yaml file to support toggling this off and on with a
the variable osbuildvm_images_build. The results will be outputted to a
subdirectory named osbuildvm-images in the out directory.

This commit also adjusts the run-make role to toggle off if target_image
is undefined or set to a falsey value. This is needed to be able to have
runs which skip the target image build and only build the osbuildvm
images.