run-make: fix toggling on/off
As part of enabling the osbuildvm-images build feature, a toggle was
added to turn off run-make when target_image is undefined/empty. However
it was a broken toggle and permanently kept run-make disabled.

This commit fixes the behaviour so it correctly toggles on/off when
target_image is undefined/empty vs defined/populated.