run-make: fix toggling on/off (2)
Try again to fix the toggling on/off (a follow up to e2e8cfc). The
fix applied there does not work when target_image is left empty. The
role throws the error: 'object of type 'NoneType' has no len()'.

This time, stick to the standard truthfulness testing and skip using
length() > 0 to test the variable.