hetzner-cloud-instantiate: don't use ping
Switch from wait_for_connection to wait_for and configure it to check
for an ssh connection from the Ansible host to determine liveliness.
This is because ping is sometimes not possible from GitHub public
runners so an alternative connection check is necessary to make asz
viable in GitHub actions as part of a CI/CD process.